refactor: 사이드 패널 안에 사이드 바 배치
